Added tracking plugin dll, unity project to run it
The dll does both tracking and pose estimation, and passes the resulting
data (ids, corner coordinates, tvecs and rvecs) to the unity side by
giving it pointers to plugin-side variables.
The variables are arrays on the heap with all the data, and get
overwritten when the next tracking call is done. The aruco module
partially deals with arrays of arrays etc, so for transfer the plugin
copies all data into flat arrays (e.g. for corners, the result is simply
a flat array of floats, representing x,y value pairs in sequence).
The C# side can use the pointers from the plugin to marshal the data
back into managed arrays from where it can handle it as it wishes.

The unity project currently executes tracking when spacebar is
pressed, then extracts all tracking and pose information and prints it
for validation. In the app, the webcam image is overlayed over the
camera, so later objects can be placed in 3D that match the marker
position and rotation, using the pose estimation values.