Started translating pose to unity object pos/rot
The aruco tracker script now places one quad for the first marker found,
which is intended to exactly match the marker position on the video
feed. The quad already matches in its position, however rotation has
some discontinuities where it is wildly off from what's visible on the
video. I am in the process of trying to fix this.`