Fix for marker rotation, debug code,aruco lib test
Found the origin of marker rotation values not matching - the camera
matrix was accidentally reversed when inputting it into the program (the
code was accessing the matrix with x,y, when it should access with y,x).
This lead to incorrect rotation values. The corrected matrix leads to
rotation matching consistently.

NOTE: This code is very messy, with many debugging functions (such as
generating the full rotation matrix from the rodrigues vector) included.
This is left here simply for housekeeping to have access to it later if
needed, but will be removed in the next commit.

There is also a version of the aruco plugin using the full Aruco library
- It does not support the rvec/tvec functionality, and will be left like
this for now since the OpenCV-only plugin is working fine. However, the
plugin may get expanded if compilation of the Aruco library for the
HoloLens does not prove to be a problem, and its functionality proves
useful.

Also added .gitignore rules to exclude the aruco build directory.