Parented marker objs to cam, custom cam display
Marker objects are now parented to the camera, so they should perfectly
map to what the camera is seeing even when it is moved (of course this
will only really make sense once it is tested on the HoloLens).

Tried out the Holographic Emulation mode, and the camera seems to get
set to some internal values for FOV, and also possibly position (maybe
the view position gets offset by some internal factor). None of this is
visible on the camera object, so I have approximated the FOV of the
emulation camera experimentally, and this can be used instead of the
actual camera FOV with a switch on the custom display script (which was
added because the UI canvas does not notice the different FOV, and also
lags behind the camera).