DLL issues on emulator fixed, small ball additions
The issue with the DLLs turned out to be missing dlls that the opencv
dlls dependent upon! (Dumb thing to get wrong, but the error reporting
certainly didn't help).
The opencv dlls for the editor are now also added to the project.
There is still an issue on the emulator now where the dll side throws a
stack overflow exception - however I have no good way to debug this,
since there is no information about the source.

On the HoloBall game, I reorganized some assets and changed the hole to
a "finish area marker", also started on adding a trigger collider to it
to later make a prefab out of it.