Investigated emulator crash, removed pixels allocs
The issue was the dll hitting the memory limit due to the GetPixels32()
functions allocating lots of memory that didn't get collected until the
app hit the memory max. Changed it to work without allocs for now (since
the test image can just be grabbed once and used repeatedly), and the
webcam mode has an option to output into a pre-allocated array. This
still needs testing on the device though.

Also made small adjustments to the image handle code such that the
colors array is actually fixed during the dll call (even though that's
probalby not relevant since the call is synchronous)