Tracking now runs on the HoloLens
Modified the plugin code slightly such that camera parameters are now
passed in from the unity side, instead of being hardcoded in the dll.
This allows for easier testing of tracking in the editor, but with
hololens camera images and parameters, and unifies the plugin code
across platforms.

Also took some images from the HoloLens and generated camera parameters
from them, however they had to be modified slightly such that the visual
camera center is at width/2 and height/2, instead of the false values
generated by the calibration. The current values seem to work well.

It should be noted that the values committed here are for 1408x792
images, which are what the HoloLens camera app generates. However, the
webcam sends 1280x720 images to the unity script, so the camera
parameters will have to be scaled for the different image size.
This will be added soon in a future commit.