Image downsize test; frustrum plane correction
Trying out a camera image downscaling in the OpenCV plugin, and the
associated performance gains. Nothing has been measured yet, but the
effect is very noticable on the actual device.
Right now, the code will still take the large 1408x792 image size, which
is wasteful if the image is later downsized anyway. This is done right
now because the camera intrinsics were made for this image size, however
with FOV information from the Locatable camera documentation page, an
intrinsics matrix can be generated without calibration using images of
the given size. (Use fov formualas to convert from HFOV to focal length,
which makes up the first to values of the intrinsics array. The other
two are typicallly just image width / 2 and image height / 2)

The frustrum plane code was also adjusted to not use camera values
anymore, since the display should actually always match the fov of a
given test image, not that of the unity camera. With this setup, pose
information for markers always matches what is shown on the image, and
also removes the z-axis adjustment in the tracker script. This should
eliminate most of the error seen when used on the actual hololens, as
the unity camera has a different FOV than the video camera, thus causing
the script to falsely adjust the z-axis position of markers.

This still remains to be verified on the actual device, but I am optimistic that it works
correctly. The other source of error I anticipate apart from this is the
mis-match between the video camera's position in space, and the unity
virtual camera which matches the wearer's eye position. Currently, they
are being equated, when really the video camera is some amount above and
in front of the wearer's eyes. It should be possible to find this offset
through the LocatableCamera API, specifically from taking a snapshot
with it and extracting the transform matrix information coming with
that. This will however have to be detected each time the app starts,
since the offset varies based on how a user wears the HoloLens.