Added reset event, renamed for consistency
Reset event was added so that components can handle their internal state
correctly on reset, this also fixed a bug where the goal force would be
applied when the ball had just been reset because it hadn't yet exited
the trigger.
Also renamed all variables to camel case for consistency with Unity.