Refactored aruco c# code to be cleaner&modular
The main interface with the dll is provided by the ArucoTracker class as
added in the previous commit. The ArucoRunner then is the c# class which
does the management of getting an image, feeding it to the DLL, and also
generating pose data in Unity data formats (i.e., a Vector3 position and
a Quaternion per marker).
The ArucoRunner gets its image and camera intrinsics from a CameraProvider, which can be
implemented in whatever way the user wishes to. The CameraProvider is an
abstract class instead of an Interface since Unity does not by default
serialize and display Interface class members.
The GeneralCamerProvider class provides the functionality of getting
either a webcam feed or a test image asset, as well as intrinsics for
the hololens or my own Macbook 15 webcam.

The entire tracking loop is powered by the ArucoUpdateRunner, which
actually invokes the initialisation and detection done by ArucoRunner.
This is separated to allow the user to invoke the tracking in other ways
than simply every update step.
Lastly, the display of the webcam feed / test image on the frustrum
plane was also separated out to the ImageDisplay component, which gets
its texture from the GeneralCameraProvider.