Fixed bug in offset runner, found offset
The offset runner was adding the offset after applying the camera
matrix, however the position is already camera-relative so this was
incorrect. With the fix the offset was working properly.
Also added debug information to be printed in the offset phrase
recogniser. Additionally changed the forwards and backwards keywords to
"to" and "from" since forwards and backwards were not being recognised
reliably.

Finally, found a manual offset for marker positions. The marker
positions are still not perfect, but they are within the expected
accurace (roughly += 0.5 cm, especially along camera z). This should be
enough for our purposes.