Combined low-pass and avg filter
This works fairly well for the most part and removes some wobble from
the average-only filtering. However, the quaternion averaging fails
horribly on quick turns since the filtering now enforces quaternions to
be further apart. It may be useful to implement checks during quaternion
averaging such that it stops once it finds a quaternion that is too far
away from the total average or the first one.