Added game scene for tracking testing
The scene is the same as the current main scene, only using a speed
boost. This is now mapped to a marker using the library. The tracked
object component has also been changed slightly such that it now places
objects in world space. This prevents persistent objects to move with
the camera and leave their place.