Added snapping test project
The snapping project is a testbed for adding the snap-to-plane
functionality. Currently objects get snapped correctly to the plane,
their rotation is aligned to the plane, and their position is snapped to
a gridsize on the plane. Snappnig rotation to 90-degrees intervals in
plane space is still left to be done.