Added freeze toggle to ArucoUpdater
Also made the GameManager toggle the freeze based on game playing/pause
state.
Also made the ArucoUpdater stop tracking if the hololens is moved to
fast to prevent markers from shifting when looking away from them.
Also added a video recorder test script to the aruco plugin test. This
was to check if the locatable camera video recording would work in
parallel to the webcam texture access. Unfortunately, it does not.